short-me-web/requirements.txt
flask
jinja2
mysql
cryptography